added button for navigation back to current date
this will be the final commit for this file, this HTML is for the 'calendarnav' Javascript file. the Javascript is where the bulk of my code is and where i spent the most time learning. now that this mini project is complete i can implement this into my NASA API project.