final commit, all working, current date btn added
this will be my final commit for this file. all buttons are working as well as the newly added 'current date' which navigates back to the current date according to the computer. i also added notes for most of the function section to describe my thinking and clearly show each lines purpose