DNA sequence final commit
this algorithm is a command line interface that allows the user to set pairs with two alphanumeric characters. they can then create a sequence by typing in string of characters. if one of the pair characters are found in the string the other will be paired along with it. this program uses the 'readline' module to get input and basic JavaScript methods to produce output. use NodeJS to run this program in the command line