final commit for missing character finder
this program works by taking user input via the 'readline' module. then it splits the string into an array of single characters. this array is then sorted which makes finding the first and last num/letter-char a lot easier. if numbers are present the 0th index will always be the smallest number and if letters are present the last index will always be the closest to the end of the alphabet. some basic logic is used to set the 4 params need, numStart/End and alphaStart/End. these params are passed to two functions for finding the missing characters. these functions are only called if their respective char type is present in the user's input. when these functions run they iterate a for-loop where i is initialized as the 'start' variable's charcode. and the loop ends when i reaches the 'end' variable's charcode. the string.fromCharCode is checked to be present in the users array. if not found that char is added to the respective 'missingArray'. finally the display function will be called to show the user what characters were missing