prime numbers calculator
this program allows the user to set a max value and then an array of all prime numbers below and including the max will be generated. then the user can select one of the four basic operations to preform on the sequence that has been created for them. finally there results will be displayed. a final answer is console logged as well as a standard math equation that shows all the numbers uses and the math operation they selected