inital commit for the steamroll array algorithm
I approached the solution to this algorithm recursively. because I want this algorithm to work no matter  how many sub-arrays are present I created a recursive function that will keep 'digging deeper' into the array until a single value can be extracted. I am going to work on creating a cleaner looking function for this as well as implement it into a command line interface so you can enter an array in string form.