flatten multi dimensional arrays
I have created two new functions that flatten arrays that are multidimensional. one uses the spread operator and a for-loop to replace sub-arrays with a 'spread' version of itself. the other uses recursion and the array method .flat to continually flatten the array until it is a single dimension array.