Compares Runtime on two functions
Ask the user how many times each function should be ran and the average times are showed to the user after the test are complete. some more work will be added to this app. I will be reusing it often to test my algorithms and see how efficient i can make them